refactor: delete batch_kwargs_anonymizer
